refactor(server): clean up re-export chain in analysis.ts
SQLiteMessageRow, InsightRow, SessionData, and ParseError were
re-exported from analysis.ts so routes could use a single import point.
This created false dependencies — changes to prompt-types.ts or
analysis-db.ts would trigger recompilation of all analysis.ts consumers
even when the route-level code was untouched.

Routes and index.ts now import types directly from their canonical
modules: SQLiteMessageRow and ParseError from prompt-types.ts,
InsightRow and SessionData from analysis-db.ts.
