+++ server/src/llm/index.ts
@@ -4,5 +4,6 @@ export type { LLMClient, LLMMessage, LLMResponse, ChatOptions } from './types.js
 export type { LLMProvider, LLMProviderConfig } from './types.js';
 export { createLLMClient, createClientFromConfig, loadLLMConfig, isLLMConfigured, testLLMConfig } from './client.js';
 export { analyzeSession, analyzePromptQuality, findRecurringInsights } from './analysis.js';
-export type { AnalysisResult, InsightRow, SessionData, RecurringInsightResult } from './analysis.js';
+export type { AnalysisResult, RecurringInsightResult } from './analysis.js';
+export type { InsightRow, SessionData } from './analysis-db.js';
 export { discoverOllamaModels } from './providers/ollama.js';
