+++ server/src/routes/route-helpers.ts
@@ -8,8 +8,9 @@ import { streamSSE } from 'hono/streaming';
 import { getDb } from '@code-insights/cli/db/client';
 import { trackEvent, captureError } from '@code-insights/cli/utils/telemetry';
 import { isLLMConfigured, loadLLMConfig } from '../llm/client.js';
-import type { SessionData, AnalysisResult, AnalysisOptions } from '../llm/analysis.js';
-import type { SQLiteMessageRow } from '../llm/analysis.js';
+import type { AnalysisResult, AnalysisOptions } from '../llm/analysis.js';
+import type { SessionData } from '../llm/analysis-db.js';
+import type { SQLiteMessageRow } from '../llm/prompt-types.js';
 
 /**
  * Load a session row for LLM analysis. Returns undefined if the session doesn't exist
